+++ alerter/src/alerter/factory/chainlink_node_alerting_factory.py
@@ -4,7 +4,7 @@ from datetime import timedelta
 from src.alerter.factory.alerting_factory import AlertingFactory
 from src.alerter.grouped_alerts_metric_code.node.chainlink_node_metric_code \
     import GroupedChainlinkNodeAlertsMetricCode as AlertsMetricCode
-from src.configs.alerts.chainlink_node import ChainlinkNodeAlertsConfig
+from src.configs.alerts.node.chainlink import ChainlinkNodeAlertsConfig
 from src.utils.configs import parse_alert_time_thresholds
 from src.utils.timing import (TimedTaskTracker, TimedTaskLimiter,
                               OccurrencesInTimePeriodTracker)